Update manual-entry-of-contributions.md on line 28 New Contribution Form is capitalized as it in the UI.
On lines 29 and 30 Contribution Form is capitalized.

On line 35, 36, and 37 Amount, Received Date, Receipt Date, and Status were capitalized.

On line 39 Personal Campaign Pages was capitalized.

On lines 58, 60, 61, 68, 69 Contribution, Membership, and Pledge Payments are capitalized as they are types of Payments in the UI.

On line 83 added . period at end of sentence.

On line 94 Batch Name and Edit are capitalized to match UI.

On line 97 Batch Entry Grid Screen is capitalized because it refers to a portion of the UI.

On line 103 Total Amount is capitalized because it refers to the item in the UI.

On line 109 Edit and Delete are capitalized because it refers to the item in the UI.

On line 149 removed the extra space between "Source** :". 

On line 146 Contribution Payments was capitalized because it refers to the item in the UI.

On lines 155 Membership Payments was capitalized because it refers to the item in the UI.

On line 199 added space between "on** Enter".

On lines 200 and 202 added . periods at ends of sentences.

On lines 221 and 223 Contribution, Membership, and Pledge Payments are capitalized as they are types of Payments in the UI.

On line 226 Contributions capitalized.

On line 234 added space between "**Status** field".

On line 254 removed "called" as it the sentence do not make sense. "Click on Fields net to called New Individual, New Household, or New Organization."

On line 257 Batch Entry Input Grid capitalized, also removed ", Generally" is unnecessary I put a . period after profiles to end the sentence.

On line 272 removed the first "''of" to make the sentence read better.

On lines 270, and 273 Contributions, Membership, and Payments  are capitalized as they are types of Payments in the UI.

On lines 283 and 284 Batch Entry Input Grid was capitalized as it is an area of the UI.

On line 290 Contributions, Pledge, and Payment  are capitalized as they are types of Payments in the UI.

On line 296 Membership and Payment  are capitalized as they are types of Payments in the UI.

On line 300 Wrote "CiviCRM" correctly not civicrm.