Update finding-and-viewing-contributions.md On line 5 changed Mont-To-Date and Year-To-Date to match the UI also cleaned up the code slightly to read display better.
On line 13 changed the period at the end to : to match line 8 Table Layout tab:

On line 19 Year-To-Date was capitalized to match the UI.

On line 32 to was changed to "You" as in each time YOU do a search.

On line 34 Contributions was capitalized because it is a UI function or search name.

On lines 40, 41, 43, 44, 49, 50 have the terms used in the Search UI capitalized. Such as Date Range, Contribution Amount, Contribution Status, Date Range, Total Amount, Number of Contributions, Average Contribution Mean and Mode. 

On line 65 Delete is capitalized as an action in the UI.

On line 84 Pay Later is capitalized to match UI.