Update accounting-integration.md any UI item terms will be capitalized.
On lines 25 and 26 Batch Name.

On line 27 Description, and Batch.

On lines 28, 33, 37, 42,   Batch.

On line 36 Total Amount.

On line 47 Accounting Batch. Added "Existing" between "...opening an EXISTING one will...".

On lines 51, 64, 65, 67, 70, 72, 75, 105, 108, 130, and 151 Batch. 

On lines 53 - 62 terms in the UI are capitalized as they are displayed. in the UI.

On lines 83, 84,90, 91, 134, and 135 Assigned Transactions.

On lines 98, 99, 110, 114, 118, 134, Close, Batch, and Export capitalized. 

On line 125 Batches by Status and Batch Name.

On line 126 User.

On line 129 Status.

On line 141 Exported.