Update invoicing.md On lines 4, 8, 10, 13, 24, 32, 44, 54, 59, and 62 Invoicing  / Invoice / and Invoices are capitalized as they are describing a UI item.
On line 10 added "and Credit Notes" capitalized Prefix, Invoices, Credit and Notes.

On line 11 "Due Date" capitalized to match UI.

On line 12 "Notes or Standard Terms and Invoice" capitalized to match UI.

On line 19 "Email and Print"capitalized to match UI.

On line 28 removed 1. as it is unnecessary and doesn't make sense you had 1. and 2. bullets above. and below new bullets start with 1., 2., 3. 

On line 32 I suggest adding Emailing Invoices:

On line 43 added spaces between "CC / BCC" to match line 42.

On line 60 the documentation references code line 10 however there is no clear numbering of the code. I opened the file following the path in line 58 and the HTML editor does not have numbered lines in the code. I rewrote the sentence to "To change the image on the Invoice you will need to modify the code see below, replacing the image source (img src)."