Update what-you-need-to-know.md On line 4 updated the instructions to reflect what is seen in screen shot.
On line 8 "New Pledge" is capitalized to match the UI. Consider updating the screen shot on line 44 as the UI looks different than what is shown in the documentation. At least it does in my Joomla version of CiviCRM.

On line 10 "Contribution Form" is capitalized as a named component in UI.

On line 11 "Contributions" capitalized as a part of the UI.

On line 15 added **Pledge by* as the UI section .

On line 23 "Made" capitalized to match UI.

On line 24 "Start" capitalized to match UI.

On line 28 "Date" capitalized to match UI.

On line 31 "Donation and Event Fee" capitalized to match UI.

On line 33 "Pledge and Campaign" capitalized to match UI.

On line 36 "Contribution Form" capitalized as a UI component.

On line 38 "Honoree" capitalized as a UI type.

On line 49 "Payments Page" is capitalized to match the UI.

On line 64 "New Pledge" is capitalized as it is a UI funtion.

On line 68 "Initial Reminder" is capitalized to match the UI.

On line 75 "Service Payments" the UI displays this as "Self-service Payments Page" So the following documentation changes were made to follow that.

On line 78 "Contribution Page, Payment Processor" were capitalized as UI and Financial types.

On line 79 "Contribution Page, and Self" was capitalized.

On line 80 "Payment Page and New Pledge" was capitalized to match UI.

On line 86 "Pledges" capitalized. Also added "Pledger" to match UI. "Name, and Email Address" capitalized to match UI.

On line 87 "Pledges and Payment Start Date" were capitalized to match UI.

On lines 88 "Pledge Status, Honoree" were capitalized to match UI.

On line 89 "Pledges, View, Edit, and Delete" were capitalized to match UI.

On line 91 added or 'Pending' as that is what the current UI says. "Status and Cancel" were capitalized to match UI.

On line 92 "Pledge" was capitalized.

On lines 98 and 99 "Pledges" were capitalized.

Also consider updating the image from line 103 there is now a Pledge Reports menu item in the UI. This is a what looks like a new feature.