Update everyday-tasks.md On line 1  "Contact" was capitalized. Added (**Search >> Find Contacts**). To show how to look up to Contacts.
On line 10 . Period added at end of sentence.

One line 13 "Pledge Payment" is capitalized because that is how it is displayed in the UI.

On line 14 "paid" was changed to "Completed" as seen in the UI.

On line 16 "Edit Scheduled Payment" was capitalized as seen in the UI.

On line 34 Changed the order of words to now be the same as the UI "New Data Entry Batch

On line 47 changed the **Pledge Payment** to **Payment Type Pledge Payment** to reflect the UI.

On line 48 removed the . Period to match all the other bullets. Added (not required) to match the other bullets that display (required field)

On line 49 "-   **Status**: the default will be “Open” " Has been replaced with **Description** bullet point as it is an option in the UI within this step. Added (not required) to match the other bullets that display (required field)

On line 50 "Items, Total and Number of Payment Items" capitalized to match UI and describe the function. 

On line 52 "Total Amount" capitalized to match the UI.

On line 55 there was a !!! Note that no longer fits CiviCRM UI. As there is no Status UI item.

On new line 55 **Edit** and **Delete** are both capitalized and bolded to match the UI Items and that they are selectable links in the UI. 

On line 77 changed the end of the sentence to read "...the contact in the new window that opens."

On line 99 removed "This is" and put (required field) to match the document as above.

On lines 100 and  102 add . period at end of sentence.