Update pledges-reports-and-analysis.md
On line 3 "Reports" is capitalized to match UI.

On line 6 the names of the report in the UI is "Pledge Detail Report" so I changed it.

On line 10 "Pledged But not Paid" capitalized to match Reports UI. 

On lines 18 - 20 "Pledge Summary Report" is not in the UI this description does not make sense "summarising (Shows up as a typo try "summarizing")  total pledge amounts per, say, city or month." What does that mean? This needs to be changed i'm really not sure how to update it but it is like word salad.