Update what-is-civievent.md
On line 5 added a "d" to "Required".

When referring to Event Types or specific Types of events within the Event Types drop down menu (Conference, Exhibition, Fundraiser, Meeting, Performance, or Workshop they will be capitalized in the documentation to match the UI. 

On lines 80, 83, 88, and 92. 

On line 89 added the link for this section in the code.

On 92 changed trainings for Workshops as Workshops are a type of event.

On line 95 added Default and capitalized "Roles"

On line 97 added Default and on line 98 capitalized 
"Role" to match UI.

On line 98 Attendee, Volunteer and line 99 Speaker capitalized added Host as a role.

On lines 102 and 105 Volunteers capitalized

On lines 109 and 110 Registered, No-show (replaced wait-listed) Attended and Cancelled were capitalized as UI Statuses. Wait-listed was replaced as it is not a Status in the UI.

On line 127 link was added for CiviContribute