Menu re-organisation
We had a conversation recently about how to restructure the docs to
1) make space for search kit and form builder
2) de-prioritise civireport (as search kit can do
more and more of this we want to focus efforts there.

As far as I can remember this is along the lines of what we discussed.

From my point of view the driver for this is that I am often put off
documenting search kit and form builder because I can't figure out
where to put them and the lack of clear findable documentation
is a major inhibitor to uptake